add nodejs_check macro

+++ nodejs-fixdep.rb
@@ -23,8 +23,10 @@ files.each do |f|
     if f.index(pkgname + "/package.json")
         # get its parent
         pkgjson = f.gsub('node_modules/' + pkgname + '/package.json','') + "package.json"
-        File.open(pkgjson,'r:UTF-8') {|f| json = JSON.parse(f.read)}
-	json["dependencies"][pkgname] = dep
-	File.open(pkgjson,'w:UTF-8') {|f| f.write JSON.pretty_generate(json) }
+	if File.exist? pkgjson
+          File.open(pkgjson,'r:UTF-8') {|f| json = JSON.parse(f.read)}
+	  json["dependencies"][pkgname] = dep
+	  File.open(pkgjson,'w:UTF-8') {|f| f.write JSON.pretty_generate(json) }
+	end
     end
 end
